Create a controllers factory where all collaborators are created and controllers are mounted afterwards
