+++ lib/cartodb/server.js
@@ -1,377 +1,55 @@
-var express = require('express');
+const express = require('express');
 const cors = require('./middleware/cors');
 const user = require('./middleware/user');
-var bodyParser = require('body-parser');
-var RedisPool = require('redis-mpool');
-var cartodbRedis = require('cartodb-redis');
-var _ = require('underscore');
+const bodyParser = require('body-parser');
+const _ = require('underscore');
 
-const AnalysisLayergroupController = require('./controllers/layergroup/analysis');
-const AttributesLayergroupController = require('./controllers/layergroup/attributes');
-const DataviewLayergroupController = require('./controllers/layergroup/dataview');
-const PreviewLayergroupController = require('./controllers/layergroup/preview');
-const TileLayergroupController = require('./controllers/layergroup/tile');
-
-const AnonymousMapController = require('./controllers/map/anonymous');
-const NamedMapController = require('./controllers/map/named');
-
-const AdminTemplateController = require('./controllers/template/admin');
-const PreviewTemplateController = require('./controllers/template/preview');
-const TileTemplateController = require('./controllers/template/tile');
-
-const AnalysesController = require('./controllers/analyses');
-
-const ServerInfoController = require('./controllers/server-info');
-
-var SurrogateKeysCache = require('./cache/surrogate_keys_cache');
-var NamedMapsCacheEntry = require('./cache/model/named_maps_entry');
-var VarnishHttpCacheBackend = require('./cache/backend/varnish_http');
-var FastlyCacheBackend = require('./cache/backend/fastly');
-
-var StatsClient = require('./stats/client');
+const StatsClient = require('./stats/client');
 const stats = require('./middleware/stats');
 
-var RendererStatsReporter = require('./stats/reporter/renderer');
-
-var windshaft = require('windshaft');
-var mapnik = windshaft.mapnik;
-
-var TemplateMaps = require('./backends/template_maps.js');
-var OverviewsMetadataApi = require('./api/overviews_metadata_api');
-var FilterStatsApi = require('./api/filter_stats_api');
-var UserLimitsApi = require('./api/user_limits_api');
-var AuthApi = require('./api/auth_api');
-var LayergroupAffectedTablesCache = require('./cache/layergroup_affected_tables');
-var NamedMapProviderCache = require('./cache/named_map_provider_cache');
-var PgQueryRunner = require('./backends/pg_query_runner');
-var PgConnection = require('./backends/pg_connection');
-
-var AnalysisBackend = require('./backends/analysis');
-const AnalysisStatusBackend = require('./backends/analysis-status');
-const DataviewBackend = require('./backends/dataview');
-
-var timeoutErrorTilePath = __dirname + '/../../assets/render-timeout-fallback.png';
-var timeoutErrorTile = require('fs').readFileSync(timeoutErrorTilePath, {encoding: null});
-
-var SqlWrapMapConfigAdapter = require('./models/mapconfig/adapter/sql-wrap-mapconfig-adapter');
-var MapConfigNamedLayersAdapter = require('./models/mapconfig/adapter/mapconfig-named-layers-adapter');
-var MapConfigBufferSizeAdapter = require('./models/mapconfig/adapter/mapconfig-buffer-size-adapter');
-var AnalysisMapConfigAdapter = require('./models/mapconfig/adapter/analysis-mapconfig-adapter');
-var MapConfigOverviewsAdapter = require('./models/mapconfig/adapter/mapconfig-overviews-adapter');
-var TurboCartoAdapter = require('./models/mapconfig/adapter/turbo-carto-adapter');
-var DataviewsWidgetsAdapter = require('./models/mapconfig/adapter/dataviews-widgets-adapter');
-var AggregationMapConfigAdapter = require('./models/mapconfig/adapter/aggregation-mapconfig-adapter');
-var MapConfigAdapter = require('./models/mapconfig/adapter');
-
-var StatsBackend = require('./backends/stats');
-
-const ResourceLocator = require('./models/resource-locator');
-const LayergroupMetadata = require('./utils/layergroup-metadata');
+const { mapnik } = require('windshaft');
 
 const lzmaMiddleware = require('./middleware/lzma');
 const errorMiddleware = require('./middleware/error-middleware');
+const syntaxError = require('./middleware/syntax-error');
 const servedByHostHeader = require('./middleware/served-by-host-header');
 
-module.exports = function(serverOptions) {
+const ControllersFactory = require('./controllers/factory');
+
+module.exports = function createServer (serverOptions) {
+    validateOptions(serverOptions);
+
     // Make stats client globally accessible
     global.statsClient = StatsClient.getInstance(serverOptions.statsd);
 
-    var redisPool = new RedisPool(_.defaults(global.environment.redis, {
-        name: 'windshaft-server',
-        unwatchOnRelease: false,
-        noReadyCheck: true
-    }));
-
-    redisPool.on('status', function(status) {
-        var keyPrefix = 'windshaft.redis-pool.' + status.name + '.db' + status.db + '.';
-        global.statsClient.gauge(keyPrefix + 'count', status.count);
-        global.statsClient.gauge(keyPrefix + 'unused', status.unused);
-        global.statsClient.gauge(keyPrefix + 'waiting', status.waiting);
-    });
-
-    var metadataBackend = cartodbRedis({pool: redisPool});
-    var pgConnection = new PgConnection(metadataBackend);
-    var pgQueryRunner = new PgQueryRunner(pgConnection);
-    var overviewsMetadataApi = new OverviewsMetadataApi(pgQueryRunner);
-    var filterStatsApi = new FilterStatsApi(pgQueryRunner);
-    var userLimitsApi = new UserLimitsApi(metadataBackend, {
-        limits: {
-            cacheOnTimeout: serverOptions.renderer.mapnik.limits.cacheOnTimeout || false,
-            render: serverOptions.renderer.mapnik.limits.render || 0,
-            rateLimitsEnabled: global.environment.enabledFeatures.rateLimitsEnabled
-        }
-    });
-
-    var templateMaps = new TemplateMaps(redisPool, {
-        max_user_templates: global.environment.maxUserTemplates
-    });
-
-    var surrogateKeysCache = new SurrogateKeysCache(surrogateKeysCacheBackends(serverOptions));
-
-    function invalidateNamedMap (owner, templateName) {
-        var startTime = Date.now();
-        surrogateKeysCache.invalidate(new NamedMapsCacheEntry(owner, templateName), function(err) {
-            var logMessage = JSON.stringify({
-                username: owner,
-                type: 'named_map_invalidation',
-                elapsed: Date.now() - startTime,
-                error: !!err ? JSON.stringify(err.message) : undefined
-            });
-            if (err) {
-                global.logger.warn(logMessage);
-            } else {
-                global.logger.info(logMessage);
-            }
-        });
-    }
-    ['update', 'delete'].forEach(function(eventType) {
-        templateMaps.on(eventType, invalidateNamedMap);
-    });
-
     serverOptions.grainstore.mapnik_version = mapnikVersion(serverOptions);
-
-    validateOptions(serverOptions);
 
     bootstrapFonts(serverOptions);
 
-    // initialize express server
-    var app = bootstrap(serverOptions);
+    const app = bootstrap(serverOptions);
 
-    var mapStore  = new windshaft.storage.MapStore({
-        pool: redisPool,
-        expire_time: serverOptions.grainstore.default_layergroup_ttl
-    });
+    app.use(bodyParser.json());
 
-    var onTileErrorStrategy;
-    if (global.environment.enabledFeatures.onTileErrorStrategy !== false) {
-        onTileErrorStrategy = function onTileErrorStrategy$TimeoutTile(err, tile, headers, stats, format, callback) {
+    app.use(servedByHostHeader());
 
-            function isRenderTimeoutError (err) {
-                return err.message === 'Render timed out';
-            }
+    app.use(stats({
+        enabled: serverOptions.useProfiler,
+        statsClient: global.statsClient
+    }));
 
-            function isDatasourceTimeoutError (err) {
-                return err.message && err.message.match(/canceling statement due to statement timeout/i);
-            }
+    app.use(lzmaMiddleware());
 
-            function isTimeoutError (err) {
-                return isRenderTimeoutError(err) || isDatasourceTimeoutError(err);
-            }
 
-            function isRasterFormat (format) {
-                return format === 'png' || format === 'jpg';
-            }
+    setupLogger(app, serverOptions);
 
-            if (isTimeoutError(err) && isRasterFormat(format)) {
-                return callback(null, timeoutErrorTile, {
-                    'Content-Type': 'image/png',
-                }, {});
-            } else {
-                return callback(err, tile, headers, stats);
-            }
-        };
-    }
+    app.use(cors());
+    app.use(user());
 
-    var rendererFactory = new windshaft.renderer.Factory({
-        onTileErrorStrategy: onTileErrorStrategy,
-        mapnik: {
-            redisPool: redisPool,
-            grainstore: serverOptions.grainstore,
-            mapnik: serverOptions.renderer.mapnik
-        },
-        http: serverOptions.renderer.http,
-        mvt: serverOptions.renderer.mvt
-    });
+    const controllers = new ControllersFactory({ serverOptions, environmentOptions: global.environment });
 
-    // initialize render cache
-    var rendererCacheOpts = _.defaults(serverOptions.renderCache || {}, {
-        ttl: 60000, // 60 seconds TTL by default
-        statsInterval: 60000 // reports stats every milliseconds defined here
-    });
-    var rendererCache = new windshaft.cache.RendererCache(rendererFactory, rendererCacheOpts);
-    var rendererStatsReporter = new RendererStatsReporter(rendererCache, rendererCacheOpts.statsInterval);
-    rendererStatsReporter.start();
+    controllers.regist(app);
 
-    var attributesBackend = new windshaft.backend.Attributes();
-    var previewBackend = new windshaft.backend.Preview(rendererCache);
-    var tileBackend = new windshaft.backend.Tile(rendererCache);
-    var mapValidatorBackend = new windshaft.backend.MapValidator(tileBackend, attributesBackend);
-    var mapBackend = new windshaft.backend.Map(rendererCache, mapStore, mapValidatorBackend);
-
-    var analysisBackend = new AnalysisBackend(metadataBackend, serverOptions.analysis);
-    const analysisStatusBackend = new AnalysisStatusBackend();
-
-    const dataviewBackend = new DataviewBackend(analysisBackend);
-
-    var statsBackend = new StatsBackend();
-
-    var layergroupAffectedTablesCache = new LayergroupAffectedTablesCache();
-    app.layergroupAffectedTablesCache = layergroupAffectedTablesCache;
-
-    var mapConfigAdapter = new MapConfigAdapter(
-        new MapConfigNamedLayersAdapter(templateMaps, pgConnection),
-        new MapConfigBufferSizeAdapter(),
-        new SqlWrapMapConfigAdapter(),
-        new DataviewsWidgetsAdapter(),
-        new AnalysisMapConfigAdapter(analysisBackend),
-        new AggregationMapConfigAdapter(pgConnection),
-        new MapConfigOverviewsAdapter(overviewsMetadataApi, filterStatsApi),
-        new TurboCartoAdapter()
-    );
-
-    var namedMapProviderCache = new NamedMapProviderCache(
-        templateMaps,
-        pgConnection,
-        metadataBackend,
-        userLimitsApi,
-        mapConfigAdapter,
-        layergroupAffectedTablesCache
-    );
-
-    ['update', 'delete'].forEach(function(eventType) {
-        templateMaps.on(eventType, namedMapProviderCache.invalidate.bind(namedMapProviderCache));
-    });
-
-    var authApi = new AuthApi(pgConnection, metadataBackend, mapStore, templateMaps);
-
-    var TablesExtentApi = require('./api/tables_extent_api');
-    var tablesExtentApi = new TablesExtentApi(pgQueryRunner);
-
-    const resourceLocator = new ResourceLocator(global.environment);
-    const layergroupMetadata = new LayergroupMetadata(resourceLocator);
-
-    var versions = getAndValidateVersions(serverOptions);
-
-    /*******************************************************************************************************************
-     * Routing
-     ******************************************************************************************************************/
-
-    const mapRouter = express.Router();
-    const templateRouter = express.Router();
-    const monitorRouter = express.Router();
-
-    new AnalysisLayergroupController(
-        analysisStatusBackend,
-        pgConnection,
-        mapStore,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        authApi,
-        surrogateKeysCache
-    ).register(mapRouter);
-
-    new AttributesLayergroupController(
-        attributesBackend,
-        pgConnection,
-        mapStore,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        authApi,
-        surrogateKeysCache
-    ).register(mapRouter);
-
-    new DataviewLayergroupController(
-        dataviewBackend,
-        pgConnection,
-        mapStore,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        authApi,
-        surrogateKeysCache
-    ).register(mapRouter);
-
-    new PreviewLayergroupController(
-        previewBackend,
-        pgConnection,
-        mapStore,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        authApi,
-        surrogateKeysCache
-    ).register(mapRouter);
-
-    new TileLayergroupController(
-        tileBackend,
-        pgConnection,
-        mapStore,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        authApi,
-        surrogateKeysCache
-    ).register(mapRouter);
-
-    new AnonymousMapController(
-        pgConnection,
-        templateMaps,
-        mapBackend,
-        metadataBackend,
-        surrogateKeysCache,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        mapConfigAdapter,
-        statsBackend,
-        authApi,
-        layergroupMetadata
-    ).register(mapRouter);
-
-    new NamedMapController(
-        pgConnection,
-        templateMaps,
-        mapBackend,
-        metadataBackend,
-        surrogateKeysCache,
-        userLimitsApi,
-        layergroupAffectedTablesCache,
-        mapConfigAdapter,
-        statsBackend,
-        authApi,
-        layergroupMetadata
-    ).register(templateRouter);
-
-    new TileTemplateController(
-        namedMapProviderCache,
-        tileBackend,
-        surrogateKeysCache,
-        pgConnection,
-        authApi,
-        userLimitsApi
-    ).register(templateRouter);
-
-    new PreviewTemplateController(
-        namedMapProviderCache,
-        previewBackend,
-        surrogateKeysCache,
-        tablesExtentApi,
-        metadataBackend,
-        pgConnection,
-        authApi,
-        userLimitsApi
-    ).register(mapRouter);
-
-    new AdminTemplateController(
-        authApi,
-        templateMaps,
-        userLimitsApi
-    ).register(templateRouter);
-
-    new AnalysesController(
-        pgConnection,
-        authApi,
-        userLimitsApi
-    ).register(mapRouter);
-
-    new ServerInfoController(versions).register(monitorRouter);
-
-    const { base_url_mapconfig: mapConfigBasePath, base_url_templated: templateBasePath } = serverOptions;
-
-    app.use(mapConfigBasePath, mapRouter);
-    app.use(templateBasePath, templateRouter);
-    app.use('/', monitorRouter);
-
-    /*******************************************************************************************************************
-     * END Routing
-     ******************************************************************************************************************/
-
+    app.use(syntaxError());
     app.use(errorMiddleware());
 
     return app;
@@ -381,45 +59,6 @@ function validateOptions(opts) {
     if (!_.isString(opts.base_url) || !_.isString(opts.base_url_mapconfig) || !_.isString(opts.base_url_templated)) {
         throw new Error("Must initialise server with: 'base_url'/'base_url_mapconfig'/'base_url_templated' URLs");
     }
-}
-
-function getAndValidateVersions(options) {
-    // jshint undef:false
-    var warn = console.warn.bind(console);
-    // jshint undef:true
-
-    var packageDefinition = require('../../package.json');
-
-    var declaredDependencies = packageDefinition.dependencies || {};
-    var installedDependenciesVersions = {
-        camshaft: require('camshaft').version,
-        grainstore: windshaft.grainstore.version(),
-        mapnik: windshaft.mapnik.versions.mapnik,
-        node_mapnik: windshaft.mapnik.version,
-        'turbo-carto': require('turbo-carto').version,
-        windshaft: windshaft.version,
-        windshaft_cartodb: packageDefinition.version
-    };
-
-    var dependenciesToValidate = ['camshaft', 'turbo-carto', 'windshaft'];
-    dependenciesToValidate.forEach(function(depName) {
-        var declaredDependencyVersion = declaredDependencies[depName];
-        var installedDependencyVersion = installedDependenciesVersions[depName];
-        if (declaredDependencyVersion !== installedDependencyVersion) {
-            warn(
-                'Dependency="%s" installed version="%s" does not match declared version="%s". Check your installation.',
-                depName, installedDependencyVersion, declaredDependencyVersion
-            );
-        }
-    });
-
-    // Be nice and warn if configured mapnik version is != installed mapnik version
-    if (mapnik.versions.mapnik !== options.grainstore.mapnik_version) {
-        warn('WARNING: detected mapnik version (' + mapnik.versions.mapnik + ')' +
-            ' != configured mapnik version (' + options.grainstore.mapnik_version + ')');
-    }
-
-    return installedDependenciesVersions;
 }
 
 function bootstrapFonts(opts) {
@@ -465,35 +104,6 @@ function bootstrap(opts) {
         return value;
     });
 
-    app.use(bodyParser.json());
-
-    app.use(servedByHostHeader());
-
-    app.use(stats({
-        enabled: opts.useProfiler,
-        statsClient: global.statsClient
-    }));
-
-    app.use(lzmaMiddleware());
-
-    // temporary measure until we upgrade to newer version expressjs so we can check err.status
-    app.use(function(err, req, res, next) {
-        if (err) {
-            if (err.name === 'SyntaxError') {
-                res.status(400).json({ errors: [err.name + ': ' + err.message] });
-            } else {
-                next(err);
-            }
-        } else {
-            next();
-        }
-    });
-
-    setupLogger(app, opts);
-
-    app.use(cors());
-    app.use(user());
-
     return app;
 }
 
@@ -511,25 +121,6 @@ function setupLogger(app, opts) {
         };
         app.use(global.log4js.connectLogger(global.log4js.getLogger(), _.defaults(loggerOpts, {level: 'info'})));
     }
-}
-
-function surrogateKeysCacheBackends(serverOptions) {
-    var cacheBackends = [];
-
-    if (serverOptions.varnish_purge_enabled) {
-        cacheBackends.push(
-            new VarnishHttpCacheBackend(serverOptions.varnish_host, serverOptions.varnish_http_port)
-        );
-    }
-
-    if (serverOptions.fastly &&
-        !!serverOptions.fastly.enabled && !!serverOptions.fastly.apiKey && !!serverOptions.fastly.serviceId) {
-        cacheBackends.push(
-            new FastlyCacheBackend(serverOptions.fastly.apiKey, serverOptions.fastly.serviceId)
-        );
-    }
-
-    return cacheBackends;
 }
 
 function mapnikVersion(opts) {
